refactor(Variant): move template impl in base header
